Clean currency price format

+++ app/src/main/java/com/lucas/currencylist/models/binance/BinanceCurrency.kt
@@ -2,5 +2,5 @@ package com.lucas.currencylist.models.binance
 
 data class BinanceCurrency(
     val name: String,
-    val price: Float,
+    val price: Double
 )
