+++ app/src/main/java/com/lucas/currencylist/models/buenbit/BuenbitCurrency.kt
@@ -6,7 +6,7 @@ data class BuenbitCurrency(
     val currency: String,
     @SerializedName("bid_currency") val bidCurrency: String,
     @SerializedName("ask_currency") val askCurrency: String,
-    val purchase_price: Float,
-    @SerializedName("selling_price") val sellingPrice: Float,
+    val purchase_price: Double,
+    @SerializedName("selling_price") val sellingPrice: Double,
     @SerializedName("market_identifier") val marketIdentifier: String
 )
